[level 1] Title: 정수 제곱근 판별, Time: 0.03 ms, Memory: 75.8 MB -BaekjoonHub

+++ 프로그래머스/lv1/12934. 정수 제곱근 판별/정수 제곱근 판별.java
@@ -1,6 +1,6 @@
 class Solution {
     public long solution(long n) {
-        if (Math.pow((int)Math.sqrt(n),2) == n) return (long )Math.pow(Math.sqrt(n)+1,2);
+        if (Math.pow((long)Math.sqrt(n),2) == n) return (long )Math.pow(Math.sqrt(n)+1,2);
         return -1;
     }
 }
